LayoutTests/platform/ios-wk1/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/scroll-to-fragid/scroll-frag-percent-encoded-expected.txt


FAIL Fragment Navigation: fragment id should be percent-decoded assert_equals: expected 200 but got 0
